LayoutTests/platform/wpe/fast/canvas/webgl/shader-with-comma-op-expected.txt

CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawArrays: attempt to access out of bounds arrays
CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawArrays: attempt to access out of bounds arrays
CONSOLE MESSAGE: WebGL: INVALID_OPERATION: drawArrays: attempt to access out of bounds arrays
Tests compiling programs that use the 'sequence' operator..

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL getError expected: NO_ERROR. Was INVALID_OPERATION : no error for using comma operator with void function
FAIL getError expected: NO_ERROR. Was INVALID_OPERATION : no error for using comma operator with non-void  function
FAIL getError expected: NO_ERROR. Was INVALID_OPERATION : no error for using comma operator with standard type
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
